chore: rename root package and project name to "orbit"
- Refactor the module name and root package from com.orbit.server to co.yapp.orbit.
- Update the group in build.gradle to co.yapp.orbit.
- Change the rootProject.name in settings.gradle from server to orbit.
- Rename ServerApplication to OrbitApplication and ServerApplicationTests to OrbitApplicationTests.

YAF-75
