+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'server'
+rootProject.name = 'orbit'
