+++ src/main/java/co/yapp/orbit/OrbitApplication.java
@@ -1,13 +1,13 @@
-package com.orbit.server;
+package co.yapp.orbit;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class ServerApplication {
+public class OrbitApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(ServerApplication.class, args);
+		SpringApplication.run(OrbitApplication.class, args);
 	}
 
 }
