+++ src/test/java/co/yapp/orbit/OrbitApplicationTests.java
@@ -1,10 +1,10 @@
-package com.orbit.server;
+package co.yapp.orbit;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class ServerApplicationTests {
+class OrbitApplicationTests {
 
 	@Test
 	void contextLoads() {
